+++ src/Pages/Exercises/ExerciseCard/ExerciseCard.module.scss
@@ -1,4 +1,4 @@
-@use '../../config/fontSize' as FontSize;
+@use '../../../config/fontSize' as FontSize;
 
 .exercise{
     width: 300px;
