+++ src/Pages/Exercises/ExerciseForm/ExerciseForm.module.scss
@@ -1,4 +1,4 @@
-@use '../../config/mixins' as mixins;
+@use '../../../config/mixins' as mixins;
 
 .form{
     font-family: 'Montserrat';
