+++ src/Pages/Exercises/Exercises.module.scss
@@ -4,19 +4,8 @@
 .page{
     @include Mixins.page;
     color: white;
-    .header{
-        @include Mixins.header;
-    }
     .body{
         display: flex;
         flex-direction: column;
-
-        #create{
-            @include Mixins.blue-button;
-            margin: 40px auto;
-        }
-    }
-    .container{
-        margin: 0 10px;
     }
 }
