+++ src/Pages/Exercises/ExercisesCreate/ExercisesCreate.jsx
@@ -1,0 +1,9 @@
+import style from './ExercisesCreate.module.scss';
+
+export default function ExercisesCreate() {
+    return (
+        <div className={style.container}>
+            
+        </div>
+    )
+}
