+++ src/Pages/Exercises/index.js
@@ -1,0 +1,9 @@
+import Exercises from './Exercises.jsx';
+import ExercisesCreate from './ExercisesCreate/ExercisesCreate';
+import ExercisesView from './ExercisesView/ExercisesView';
+
+export {
+    Exercises,
+    ExercisesCreate,
+    ExercisesView
+}
